wontyoubemyneighbor/agentic/requirements.txt
# LLM Provider Dependencies
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.2.0

# REST API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Async support
aiohttp>=3.9.0

# Optional: Local LLM support
# llama-cpp-python>=0.2.0
# transformers>=4.35.0
